api-audiofind-whisper/requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.29.0
python-multipart==0.0.9
tqdm==4.67.1
openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc1e8709e37c2ad90c039a4eb72a2

